fish: Functions to package and deploy a kernel
package-kernel can be called in the root of a kernel source tree and
will try to install the kernel and modules into a temporary directory
and build a compressed tarball that can be extracted to the target
system.

deploy-kernel is an unfinished function that aims to deploy a compressed
tarball and setup symlinks in /boot to be able to boot the new kernel.

Signed-off-by: Fredrik Lönnegren <fredrik@frelon.se>